sim/requirements.txt
pynacl>=1.5.0
pycryptodome>=3.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
